Consolidate manifest image helpers into image-utils
Fold mrt-asset-helper.ts, image-path-resolver.ts and asset-dimensions.ts into a
single image-utils.ts, and move getImageDimensions there from the editor provider
so all image handling lives in one module.

Drop the MrtResolution.variants field, which was built in production but only ever
read by tests, and collapse the three-function variant ranking into one comparator
with identical ordering.

Copilot-Session: 607fbdd4-8e6c-4c50-a5d7-2c9d0556fb50
